Categorical Crossentropy Loss.
Implement `class_weight`: A list of weight values, each is a float, to be applied to        the losses of the classes. If only one value is supplied, all the losses are uniformly scaled with this value. If more than one value are provided, the number of values must match the number of classes, which is the target's dimension on the input `axis`, because a mismatch will cause an `Incompatible shapes` error.

Note, if the input `axis` is not -1, the class balancing operation does not support tensors of unknown dimensions.

PiperOrigin-RevId: 411150028